move duplicated code to controller common package
